+++ Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 let package = Package(
     name: "BugfenderPackage",
     platforms: [
-        .iOS("10")
+        .iOS("11")
     ],
     products: [
         .library(
@@ -15,8 +15,8 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "BugfenderSDK",
-            url: "https://github.com/bugfender/BugfenderSDK-iOS/releases/download/1.11.0/BugfenderSDK.xcframework.zip",
-            checksum: "8683af1363a023f3217a70ca4ca826d8d32e511a223c79da9dd79eab74b33845"
+            url: "https://github.com/bugfender/BugfenderSDK-iOS/releases/download/1.12.0/BugfenderSDK.xcframework.zip",
+            checksum: "3af060b50c0e0441362429cb739b30766eb8f3fcc24df722e3d3e2311050f8da"
         )
     ]
 )